docker/ledger-gateway/requirements.txt
kafka-python>=2.0.2
paho-mqtt>=2.0.0
